+++ src/mgr/ActivePyModule.cc
@@ -11,17 +11,11 @@
  * Foundation.  See file COPYING.
  */
 
-#include "BaseMgrModule.h"
-
 #include "PyFormatter.h"
 
 #include "common/debug.h"
 
 #include "ActivePyModule.h"
-
-//XXX courtesy of http://stackoverflow.com/questions/1418015/how-to-get-python-exception-text
-#include <boost/python.hpp>
-#include "include/assert.h"  // boost clobbers this
 
 
 #define dout_context g_ceph_context
@@ -29,50 +23,27 @@
 #undef dout_prefix
 #define dout_prefix *_dout << "mgr " << __func__ << " "
 
-// decode a Python exception into a string
-std::string handle_pyerror()
-{
-    using namespace boost::python;
-    using namespace boost;
-
-    PyObject *exc, *val, *tb;
-    object formatted_list, formatted;
-    PyErr_Fetch(&exc, &val, &tb);
-    handle<> hexc(exc), hval(allow_null(val)), htb(allow_null(tb));
-    object traceback(import("traceback"));
-    if (!tb) {
-        object format_exception_only(traceback.attr("format_exception_only"));
-        formatted_list = format_exception_only(hexc, hval);
-    } else {
-        object format_exception(traceback.attr("format_exception"));
-        formatted_list = format_exception(hexc,hval, htb);
-    }
-    formatted = str("").join(formatted_list);
-    return extract<std::string>(formatted);
-}
-
 int ActivePyModule::load(ActivePyModules *py_modules)
 {
   assert(py_modules);
-  Gil gil(pMyThreadState, true);
+  Gil gil(py_module->pMyThreadState, true);
 
   // We tell the module how we name it, so that it can be consistent
   // with us in logging etc.
   auto pThisPtr = PyCapsule_New(this, nullptr, nullptr);
   auto pPyModules = PyCapsule_New(py_modules, nullptr, nullptr);
-  auto pModuleName = PyString_FromString(module_name.c_str());
+  auto pModuleName = PyString_FromString(get_name().c_str());
   auto pArgs = PyTuple_Pack(3, pModuleName, pPyModules, pThisPtr);
 
-  pClassInstance = PyObject_CallObject(pClass, pArgs);
-  Py_DECREF(pClass);
+  pClassInstance = PyObject_CallObject(py_module->pClass, pArgs);
   Py_DECREF(pModuleName);
   Py_DECREF(pArgs);
   if (pClassInstance == nullptr) {
-    derr << "Failed to construct class in '" << module_name << "'" << dendl;
+    derr << "Failed to construct class in '" << get_name() << "'" << dendl;
     derr << handle_pyerror() << dendl;
     return -EINVAL;
   } else {
-    dout(1) << "Constructed class from module: " << module_name << dendl;
+    dout(1) << "Constructed class from module: " << get_name() << dendl;
   }
 
   return load_commands();
@@ -82,7 +53,7 @@ void ActivePyModule::notify(const std::string &notify_type, const std::string &n
 {
   assert(pClassInstance != nullptr);
 
-  Gil gil(pMyThreadState, true);
+  Gil gil(py_module->pMyThreadState, true);
 
   // Execute
   auto pValue = PyObject_CallMethod(pClassInstance,
@@ -92,7 +63,7 @@ void ActivePyModule::notify(const std::string &notify_type, const std::string &n
   if (pValue != NULL) {
     Py_DECREF(pValue);
   } else {
-    derr << module_name << ".notify:" << dendl;
+    derr << get_name() << ".notify:" << dendl;
     derr << handle_pyerror() << dendl;
     // FIXME: callers can't be expected to handle a python module
     // that has spontaneously broken, but Mgr() should provide
@@ -105,7 +76,7 @@ void ActivePyModule::notify_clog(const LogEntry &log_entry)
 {
   assert(pClassInstance != nullptr);
 
-  Gil gil(pMyThreadState, true);
+  Gil gil(py_module->pMyThreadState, true);
 
   // Construct python-ized LogEntry
   PyFormatter f;
@@ -120,7 +91,7 @@ void ActivePyModule::notify_clog(const LogEntry &log_entry)
   if (pValue != NULL) {
     Py_DECREF(pValue);
   } else {
-    derr << module_name << ".notify_clog:" << dendl;
+    derr << get_name() << ".notify_clog:" << dendl;
     derr << handle_pyerror() << dendl;
     // FIXME: callers can't be expected to handle a python module
     // that has spontaneously broken, but Mgr() should provide
@@ -187,7 +158,7 @@ int ActivePyModule::handle_command(
   assert(ss != nullptr);
   assert(ds != nullptr);
 
-  Gil gil(pMyThreadState, true);
+  Gil gil(py_module->pMyThreadState, true);
 
   PyFormatter f;
   cmdmap_dump(cmdmap, &f);
