+++ src/mgr/ActivePyModule.h
@@ -58,11 +58,9 @@ private:
   std::string uri;
 
 public:
-  ActivePyModule(const std::string &module_name_,
-      PyObject *pClass_,
-      const SafeThreadState &my_ts_,
+  ActivePyModule(PyModuleRef py_module_,
       LogChannelRef clog_)
-    : PyModuleRunner(module_name_, pClass_, my_ts_, clog_)
+    : PyModuleRunner(py_module_, clog_)
   {}
 
   int load(ActivePyModules *py_modules);
